Homework/print cargo (#28)
* Working version

* 1st step refactor

* Small fix

* align travelTests

* buy moved

* buy renamed

* sell renamed

* header files renamed

* travel name

* Travel.cpp renamed

* Splitting execute for cargo
