+++ shm/include/command.hpp
