+++ shm/tests/TravelTests.cpp
@@ -1,9 +1,9 @@
-#include "Command.hpp"
-#include "Travel.hpp"
+#include "command.hpp"
 #include "gtest/gtest.h"
 #include "locators.hpp"
 #include "map.hpp"
 #include "player.hpp"
+#include "travel.hpp"
 
 class TravelCommandTest : public ::testing::Test {
 public:
